requirements.txt
fastapi
uvicorn
pydantic
asyncpg
httpx
python-telegram-bot
gputil
psutil
